requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
bcrypt==4.0.1
pyotp==2.9.0
segno==1.6.1
itsdangerous==2.1.2
